Rename ambiguous fragments
