+++ app/src/main/java/com/marverenic/music/data/inject/JockeyGraph.java
@@ -9,14 +9,17 @@ import com.marverenic.music.ui.common.playlist.AppendPlaylistDialogFragment;
 import com.marverenic.music.ui.common.playlist.CreatePlaylistDialogFragment;
 import com.marverenic.music.ui.common.playlist.PlaylistCollisionDialogFragment;
 import com.marverenic.music.ui.library.AlbumListFragment;
-import com.marverenic.music.ui.library.ArtistFragment;
+import com.marverenic.music.ui.library.ArtistListFragment;
 import com.marverenic.music.ui.library.LibraryActivity;
 import com.marverenic.music.ui.library.LibraryFragment;
-import com.marverenic.music.ui.library.PlaylistFragment;
+import com.marverenic.music.ui.library.PlaylistListFragment;
 import com.marverenic.music.ui.library.SongFragment;
 import com.marverenic.music.ui.library.album.AlbumFragment;
+import com.marverenic.music.ui.library.artist.ArtistFragment;
+import com.marverenic.music.ui.library.genre.GenreFragment;
+import com.marverenic.music.ui.library.playlist.PlaylistFragment;
 import com.marverenic.music.ui.library.playlist.edit.AutoPlaylistEditFragment;
-import com.marverenic.music.ui.nowplaying.GenreFragment;
+import com.marverenic.music.ui.library.GenreListFragment;
 import com.marverenic.music.ui.nowplaying.MiniplayerFragment;
 import com.marverenic.music.ui.nowplaying.NowPlayingFragment;
 import com.marverenic.music.ui.nowplaying.PlayerControllerFragment;
@@ -44,18 +47,18 @@ public interface JockeyGraph {
     void inject(SongFragment fragment);
     void inject(AlbumListFragment fragment);
     void inject(AlbumFragment fragment);
-    void inject(ArtistFragment fragment);
-    void inject(PlaylistFragment fragment);
-    void inject(GenreFragment fragment);
+    void inject(ArtistListFragment fragment);
+    void inject(PlaylistListFragment fragment);
+    void inject(GenreListFragment fragment);
     void inject(NowPlayingFragment fragment);
     void inject(PlayerControllerFragment fragment);
     void inject(QueueFragment fragment);
     void inject(EqualizerFragment fragment);
     void inject(PreferenceFragment fragment);
     void inject(DirectoryListFragment fragment);
-    void inject(com.marverenic.music.ui.library.artist.ArtistFragment fragment);
-    void inject(com.marverenic.music.ui.library.playlist.PlaylistFragment playlistFragment);
-    void inject(com.marverenic.music.ui.library.genre.GenreFragment fragment);
+    void inject(ArtistFragment fragment);
+    void inject(PlaylistFragment playlistFragment);
+    void inject(GenreFragment fragment);
     void inject(AutoPlaylistEditFragment fragment);
 
     void inject(MiniplayerFragment fragment);
